Black box tests for use cases.
6 black box tests for use cases. Describing user interactions with the interface and ensuring full functionality.
